Move running statistics to recipes
