+++ more_itertools/more.py
@@ -3,11 +3,9 @@ import math
 from collections import Counter, defaultdict, deque
 from collections.abc import Sequence
 from contextlib import suppress
-from dataclasses import dataclass
 from functools import cached_property, partial, wraps
 from heapq import heapify, heapreplace
 from itertools import (
-    accumulate,
     chain,
     combinations,
     compress,
@@ -50,8 +48,6 @@ from .recipes import (
     is_prime,
     nth,
     powerset,
-    running_mean,
-    running_median,
     sieve,
     take,
     unique_everseen,
@@ -62,7 +58,6 @@ from .recipes import (
 __all__ = [
     'AbortThread',
     'SequenceView',
-    'Stats',
     'adjacent',
     'all_unique',
     'always_iterable',
@@ -148,9 +143,6 @@ __all__ = [
     'rlocate',
     'rstrip',
     'run_length',
-    'running_min',
-    'running_max',
-    'running_statistics',
     'sample',
     'seekable',
     'serialize',
@@ -5463,121 +5455,3 @@ class _concurrent_tee:
                     link[1] = [None, None]
         value, self.link = link
         return value
-
-
-def _windowed_running_min(iterator, maxlen):
-    sis = deque()  # Strictly increasing subsequence
-    for index, value in enumerate(iterator):
-        if sis and sis[0][0] == index - maxlen:
-            sis.popleft()
-        while sis and not sis[-1][1] < value:  # Remove non-increasing values
-            sis.pop()
-        sis.append((index, value))  # Most recent value at position -1
-        yield sis[0][1]  # Window minimum at position 0
-
-
-def running_min(iterable, *, maxlen=None):
-    """Smallest of values seen so far or values in a sliding window.
-
-    Set *maxlen* to a positive integer to specify the maximum size
-    of the sliding window.  The default of *None* is equivalent to
-    an unbounded window.
-
-    For example:
-
-        >>> list(running_min([4, 3, 7, 0, 8, 1, 6, 2, 9, 5]))
-        [4, 3, 3, 0, 0, 0, 0, 0, 0, 0]
-
-        >>> list(running_min([4, 3, 7, 0, 8, 1, 6, 2, 9, 5], maxlen=3))
-        [4, 3, 3, 0, 0, 0, 1, 1, 2, 2]
-
-    Supports numeric types such as int, float, Decimal, and Fraction,
-    but not complex numbers which are unorderable.
-    """
-
-    iterator = iter(iterable)
-
-    if maxlen is None:
-        return accumulate(iterator, func=min)
-
-    if maxlen <= 0:
-        raise ValueError('Window size should be positive')
-
-    return _windowed_running_min(iterator, maxlen)
-
-
-def _windowed_running_max(iterator, maxlen):
-    sds = deque()  # Strictly decreasing subsequence
-    for index, value in enumerate(iterator):
-        if sds and sds[0][0] == index - maxlen:
-            sds.popleft()
-        while sds and not sds[-1][1] > value:  # Remove non-decreasing values
-            sds.pop()
-        sds.append((index, value))  # Most recent value at position -1
-        yield sds[0][1]  # Window maximum at position 0
-
-
-def running_max(iterable, *, maxlen=None):
-    """Largest of values seen so far or values in a sliding window.
-
-    Set *maxlen* to a positive integer to specify the maximum size
-    of the sliding window.  The default of *None* is equivalent to
-    an unbounded window.
-
-    For example:
-
-        >>> list(running_max([4, 3, 7, 0, 8, 1, 6, 2, 9, 5]))
-        [4, 4, 7, 7, 8, 8, 8, 8, 9, 9]
-
-        >>> list(running_max([4, 3, 7, 0, 8, 1, 6, 2, 9, 5], maxlen=3))
-        [4, 4, 7, 7, 8, 8, 8, 6, 9, 9]
-
-    Supports numeric types such as int, float, Decimal, and Fraction,
-    but not complex numbers which are unorderable.
-    """
-
-    iterator = iter(iterable)
-
-    if maxlen is None:
-        return accumulate(iterator, func=max)
-
-    if maxlen <= 0:
-        raise ValueError('Window size should be positive')
-
-    return _windowed_running_max(iterator, maxlen)
-
-
-@dataclass(frozen=True, slots=True)
-class Stats:
-    size: int
-    minimum: float
-    median: float
-    maximum: float
-    mean: float
-
-
-def running_statistics(iterable, *, maxlen=None):
-    """Statistics for values seen so far or values in a sliding window.
-
-    Set *maxlen* to a positive integer to specify the maximum size
-    of the sliding window.  The default of *None* is equivalent to
-    an unbounded window.
-
-    Yields instances of a ``Stats`` dataclass with fields for the dataset *size*,
-    *minimum* value, *median* value, *maximum* value, and the arithmetic *mean*.
-
-    Supports numeric types such as int, float, Decimal, and Fraction,
-    but not complex numbers which are unorderable.
-    """
-
-    # fmt: off
-    t0, t1, t2, t3 = tee(iterable, 4)
-    return map(
-        Stats,
-        count(1) if maxlen is None else chain(range(1, maxlen), repeat(maxlen)),
-        running_min(t0, maxlen=maxlen),
-        running_median(t1, maxlen=maxlen),
-        running_max(t2, maxlen=maxlen),
-        running_mean(t3, maxlen=maxlen),
-    )
-    # fmt: on
